+++ src/main/java/com/bitmart/api/request/contract/prv/OrderRequest.java
@@ -4,7 +4,6 @@ import com.bitmart.api.annotations.ParamKey;
 import com.bitmart.api.request.Auth;
 import com.bitmart.api.request.CloudRequest;
 import com.bitmart.api.request.Method;
-import com.google.gson.annotations.SerializedName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -16,12 +15,17 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class OrderRequest extends CloudRequest {
 
-    @ParamKey("symbol")
-    @SerializedName("symbol")
-    private String symbol;     //Contract Trading pair: symbol (Optional, return the market information of all trading pairs by default)
+    /**
+     * Symbol of the contract(like BTCUSDT)
+     */
+    @ParamKey(value = "symbol", required = true)
+    private String symbol;
 
-    @ParamKey("order_id")
-    private String order_id;
+    /**
+     * Order ID
+     */
+    @ParamKey(value = "order_id", required = true)
+    private String orderId;
 
     /**
      * url: GET https://api-cloud.bitmart.com/contract/private/order