+++ src/main/java/com/bitmart/api/request/contract/pub/DetailsRequest.java
@@ -4,7 +4,6 @@ import com.bitmart.api.annotations.ParamKey;
 import com.bitmart.api.request.Auth;
 import com.bitmart.api.request.CloudRequest;
 import com.bitmart.api.request.Method;
-import com.google.gson.annotations.SerializedName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -16,9 +15,12 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class DetailsRequest extends CloudRequest {
 
+    /**
+     * Symbol of the contract(like BTCUSDT)
+     *   (Optional, return the market information of all trading pairs by default)
+     */
     @ParamKey("symbol")
-    @SerializedName("symbol")
-    private String symbol;     //Contract Trading pair: symbol (Optional, return the market information of all trading pairs by default)
+    private String symbol;
 
     /**
      * url: GET https://api-cloud.bitmart.com/contract/public/details