+++ src/main/java/com/bitmart/api/request/contract/pub/KlineRequest.java
@@ -4,7 +4,6 @@ import com.bitmart.api.annotations.ParamKey;
 import com.bitmart.api.request.Auth;
 import com.bitmart.api.request.CloudRequest;
 import com.bitmart.api.request.Method;
-import com.google.gson.annotations.SerializedName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -16,18 +15,29 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class KlineRequest extends CloudRequest {
 
-    @ParamKey("symbol")
-    @SerializedName("symbol")
-    private String symbol;     //Contract Trading pair: symbol (Optional, return the market information of all trading pairs by default)
+    /**
+     * Symbol of the contract(like BTCUSDT)
+     */
+    @ParamKey(value = "symbol", required = true)
+    private String symbol;
 
+    /**
+     * K-Line step, default is 1 minute. step: 1, 3, 5, 15, 30, 60, 120, 240, 360, 720, 1440, 4320, 10080
+     */
     @ParamKey("step")
-    private Long step;          //step
+    private Integer step;
 
-    @ParamKey("start_time")
-    private Long start_time;    //Start timestamp (in seconds, UTC+0 TimeZome)
+    /**
+     * Start time. Timestamps need to be in seconds
+     */
+    @ParamKey(value = "start_time", required = true)
+    private Long startTime;
 
-    @ParamKey("end_time")
-    private Long end_time;      //End timestamp (in seconds, UTC+0 TimeZome)
+    /**
+     * End time. Timestamps need to be in seconds
+     */
+    @ParamKey(value = "end_time", required = true)
+    private Long endTime;
 
     /**
      * url: GET https://api-cloud.bitmart.com/contract/public/kline