+++ src/main/java/com/bitmart/api/request/contract/pub/OpenInterestRequest.java
@@ -4,7 +4,6 @@ import com.bitmart.api.annotations.ParamKey;
 import com.bitmart.api.request.Auth;
 import com.bitmart.api.request.CloudRequest;
 import com.bitmart.api.request.Method;
-import com.google.gson.annotations.SerializedName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -16,12 +15,14 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class OpenInterestRequest extends CloudRequest {
 
-    @ParamKey("symbol")
-    @SerializedName("symbol")
-    private String symbol;     //Contract Trading pair: symbol (Optional, return the market information of all trading pairs by default)
+    /**
+     * Symbol of the contract(like BTCUSDT)
+     */
+    @ParamKey(value = "symbol", required = true)
+    private String symbol;
 
     /**
-     * url: GET https://api-cloud.bitmart.com/contract/public/open-interest
+     * url: GET <a href="https://api-cloud.bitmart.com/contract/public/open-interest">...</a>
      * Applicable for querying the open interest and open interest value data of the specified contract
      */
     public OpenInterestRequest() {