+++ src/main/java/com/bitmart/api/request/spot/pub/StepsRequest.java
@@ -14,6 +14,7 @@ import lombok.experimental.Accessors;
 @Data
 @ToString
 @Accessors(chain = true)
+@Deprecated
 public final class StepsRequest extends CloudRequest {
 
     /**