+++ src/test/java/com/bitmart/api/TestSystem.java
@@ -2,8 +2,6 @@ package com.bitmart.api;
 
 import com.bitmart.api.common.CloudException;
 import com.bitmart.api.common.CloudResponse;
-import com.bitmart.api.common.GlobalConst;
-import com.bitmart.api.key.CloudKey;
 import com.bitmart.api.request.system.pub.SystemServiceRequest;
 import com.bitmart.api.request.system.pub.SystemTimeRequest;
 import org.junit.jupiter.api.DisplayName;
@@ -14,14 +12,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 final class TestSystem {
 
-    private static final String CLOUD_URL = GlobalConst.CLOUD_URL;
-    private static final String API_KEY = "";
-    private static final String API_SECRET = "";
-    private static final String API_MEMO = "";
     private static Call call;
 
     TestSystem(){
-        CloudContext cloudContext = new CloudContext(CLOUD_URL, new CloudKey(API_KEY, API_SECRET, API_MEMO));
+        CloudContext cloudContext = new CloudContext();
         call = new Call(cloudContext);
     }